backend/requirements.txt
asgiref==3.5.2
azure-core==1.34.0
azure-storage-blob==12.25.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
diff-match-patch==20241021
Django==4.1.3
django-extensions==3.2.1
django-filter==22.1
django-import-export==4.1.1
django-modeltranslation==0.18.11
django-storages==1.14.6
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
django-modeladmin-reorder>=0.3.1
idna==3.10
inflection==0.5.1
isodate==0.7.2
pillow==11.2.1
psycopg2-binary==2.9.9
pycparser==2.22
python-dotenv==0.21.0
pytz==2022.6
requests==2.32.4
six==1.17.0
sqlparse==0.4.3
tablib==3.5.0
typing_extensions==4.14.0
tzdata==2022.6
urllib3==2.5.0
whitenoise[brotli]==6.7.0
sentry-sdk>=1.24.0